Add STOP register for stopping traffic via SPI.
This commit adds a new writable register to the HSS multiplexer register bank,
'STOP', which allows each of the 8 outgoing packet flows for the HSS stream to
be blocked off individually. This change allows SpiNNaker machines to be
soft-partitioned at board boundaries by setting the STOP bits accordingly. Note
that the recieve side is not blocked: it is assumed that both sides of a link
will have their transmitters STOPped to facilitate isolation.

To avoid adding another pipeline stage, this feature is implemented by allowing
the buffer-full signal to be forced by the STOP register value. As a result,
channels which have been STOPped will report their TX buffers being full.