Obey rdy/vld protocol.
Modifies the implementation to ensure that the ready signal is not deasserted
until a valid packet has been accepted. This (unfortunately) means that the
first packet arriving after 'stop' is set to 1 will still be sent.

Unfortunately this is not sufficient to ensure reliable booting in the general
case since the first packet sent during boot by the boot rom code irreversably
changes the state of every chip it encounters.