Stop packet transmission at frame_tx level.
This commit changes the implementation of the STOP register. The register is
now a single bit and control is now at the granularity of *all* data flowing
out over a given link.

This alternative implementation prevents data frames from being sent down the
wire whenever the stop register is 1 and allows it otherwise.